refactor events table
